chore: convert main.js to renderer.ts
